43- Configure --time=ftime: wrong reported timings
fixes an overflow. Previous commit 38f0d1df fixed the test suite

+++ src/language/init.c
@@ -2124,7 +2124,7 @@ timer_start(pari_timer *T)
 {
   struct timeb t;
   ftime(&t);
-  T->us = t.millitm * 1000;
+  T->us = ((long)t.millitm) * 1000;
   T->s  = t.time;
 }
